Refactored rebalance into market_state
